pyproject.toml
[tool.poetry]
name = "fastapi1"
version = "0.1.0"
description = "A FastAPI project"
authors = [
    "Mehwash Shahzadi <mehwashshahzadi@gmail.com>"
]
readme = "README.md"
packages = [{ include = "fastapi1" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlmodel = "^0.0.24"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

